Separate binding.cpp for pybind functionality; further organization of definitions into headers; test Makefile
