+++ pennylane_lightning/src/binding.cpp
@@ -1,0 +1,29 @@
+// Copyright 2020 Xanadu Quantum Technologies Inc.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+//     http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+/**
+ * @file
+ * Contains the main `apply()` function for applying a set of operations to a multiqubit
+ * statevector.
+ *
+ * Also includes PyBind boilerplate for interfacing with Python.
+ */
+#include "pybind11/stl.h"
+#include "pybind11/eigen.h"
+#include "lightning_qubit.hpp"
+
+PYBIND11_MODULE(lightning_qubit_ops, m)
+{
+    m.doc() = "lightning.qubit apply() method using Eigen";
+    m.def("apply", apply, "lightning.qubit apply() method");
+}
